61853 model- added check for negative parameters
Added a check in case the Gauss-Markov interpolation method calculates
negative parameters when interpolating for the 61853 module model. This
was happening frequently with the diode reverse saturation current (Io)
parameter, likely due to the high variation in magnitudes of the
parameter across the temperature-irradiance space (one example case goes
from E-5 to E-16 across the space). Multiple different scaling methods
did not resolve the issue, so it cannot be simply from the small
magnitudes of this parameter. Therefore, if the GM interpolation
calculates a negative value for any of the 5 parameters, the model will
not use that value. If the nearest value is sufficiently close, it will
simply use the nearest value for the parameter, otherwise it will
default to the normal 5-parameter model value at STC. This improvement
still needs to be tested against the measured data.